requirements.txt
ccxt>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0
scikit-learn>=1.0.0
python-telegram-bot>=13.0
rich>=10.0.0
pyyaml>=6.0.0
joblib>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.26.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
